requirements.txt
google-genai
edge-tts
openai-whisper
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
pydantic
tavily-python
